+++ src/main/groovy/com/bmuschko/gradle/docker/utils/ThreadContextClassLoader.groovy
@@ -64,6 +64,33 @@ interface ThreadContextClassLoader {
     def createVolumes(List<Object> volumes)
 
     /**
+     * Creates instance of <a href="https://github.com/docker-java/docker-java/blob/master/src/main/java/com/github/dockerjava/api/model/Link.java">Link</a>
+     * from thread context classloader.
+     *
+     * @param link a container link
+     * @return Instance
+     */
+    def createLink(String link)
+
+    /**
+     * Creates instance of <a href="https://github.com/docker-java/docker-java/blob/master/src/main/java/com/github/dockerjava/api/model/Links.java">Links</a>
+     * from thread context classloader.
+     *
+     * @param volumes List of Links
+     * @return Instance
+     */
+    def createLinks(List<Object> links)
+
+    /**
+     * Creates instance of <a href="https://github.com/docker-java/docker-java/blob/master/src/main/java/com/github/dockerjava/api/model/HostConfig.java">Links</a>
+     * from thread context classloader.
+     *
+     * @param hostConfigProperties a map containing all HostConfig properties to be set. The map entry key is the property name while the map entry value is the property value.
+     * @return Instance
+     */
+    def createHostConfig(Map<String, String> hostConfigProperties)
+
+    /**
      * Creates instance of <a href="https://github.com/docker-java/docker-java/blob/master/src/main/java/com/github/dockerjava/api/model/InternetProtocol.java">InternetProtocol</a>
      * from thread context classloader.
      *