+++ RELEASE_NOTES.md
@@ -1,3 +1,14 @@
+### Version 2.3 (April 18, 2015)
+
+* Added task `DockerInspectContainer` - [Pull request 44](https://github.com/bmuschko/gradle-docker-plugin/pull/44).
+* Added property `containerName` to task `DockerCreateContainer` - [Pull request 44](https://github.com/bmuschko/gradle-docker-plugin/pull/44).
+* Allow for linking containers for task `DockerCreateContainer` - [Pull request 53](https://github.com/bmuschko/gradle-docker-plugin/pull/53).
+* Usage of [docker-java library](https://github.com/docker-java/docker-java) version 1.2.0.
+
+### Version 2.2 (April 12, 2015)
+
+* Usage of [docker-java library](https://github.com/docker-java/docker-java) version 1.1.0.
+
 ### Version 2.1 (March 24, 2015)
 
 * Renamed property `registry` to `registryCredentials` for plugin extension and tasks implementing `RegistryCredentialsAware` to better indicate its purpose.